+++ app/quizzes/writer/writer-form.tsx
@@ -1,16 +1,13 @@
 'use client';
 
-import { useRef } from 'react';
 import Button from '@/components/common/buttons/button';
-import { Button as ShadcnButton } from '@/components/ui/button';
 import MarkdownEditor from '@/components/common/markdown/markdown-editor';
 import { Label } from '@/components/ui/label';
 import { Input } from '@/components/ui/input';
-import { Badge } from '@/components/ui/badge';
-import { Cross1Icon, PlusIcon } from '@radix-ui/react-icons';
-import { type SubmitHandler, useForm, useFieldArray } from 'react-hook-form';
+import { type SubmitHandler, useForm } from 'react-hook-form';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { z } from 'zod';
+import Hints from './hints';
 
 // 제어 컴포넌트여야되는 거
 // 1. 힌트 (가변 갯수)
@@ -34,7 +31,7 @@ const formSchema = z.object({
   answer: z.string().min(1).max(100),
 });
 
-type Inputs = z.infer<typeof formSchema>;
+export type Inputs = z.infer<typeof formSchema>;
 
 export default function WriterForm() {
   const form = useForm<Inputs>({
@@ -98,74 +95,5 @@ export default function WriterForm() {
         작성 완료
       </Button>
     </form>
-  );
-}
-
-function Hints({ form }: { form: ReturnType<typeof useForm<Inputs>> }) {
-  const { control } = form;
-
-  const {
-    append,
-    fields: hints,
-    remove,
-  } = useFieldArray({ control, name: 'hints' });
-
-  const inputRef = useRef<HTMLInputElement>(null);
-
-  const addHint = () => {
-    if (!inputRef.current || !inputRef.current.value) {
-      return;
-    }
-
-    const value = inputRef.current.value;
-    inputRef.current.value = '';
-
-    if (hints.some((hint) => hint.value === value)) {
-      return;
-    }
-
-    append({ value });
-  };
-
-  const removeHint = (id: string) => {
-    remove(hints.findIndex((hint) => hint.id === id));
-  };
-
-  return (
-    <>
-      <div className="flex w-full space-x-2">
-        <Input
-          id="hint"
-          autoComplete="off"
-          ref={inputRef}
-          onKeyDown={(e) => e.key === 'Enter' && addHint()}
-        />
-        <ShadcnButton
-          type="button"
-          variant="secondary"
-          size="icon"
-          onClick={() => addHint()}
-        >
-          <PlusIcon />
-        </ShadcnButton>
-      </div>
-      <div className="mt-2 flex gap-2">
-        {hints.map((hint, idx) => (
-          <Badge
-            key={idx}
-            className="flex items-center gap-1"
-            variant="secondary"
-          >
-            {hint.value}
-            <Cross1Icon
-              className="cursor-pointer"
-              width={10}
-              height={10}
-              onClick={() => removeHint(hint.id)}
-            />
-          </Badge>
-        ))}
-      </div>
-    </>
   );
 }
